refact: OrderPage decomposed down to several components
